requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7
pydantic-settings>=2.3
driftpy>=0.8.82
lighter-python @ git+https://github.com/elliottech/lighter-python.git
